# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductDetailedSpecs = { [key: string]: unknown };

export interface Product {
  id: string;
  /** @nullable */
  gtin?: string | null;
  title: string;
  detailedSpecs: ProductDetailedSpecs;
  /** @minimum 0 */
  inventoryCount: number;
  /** @minimum 0 */
  humanPrice: number;
  /** @minimum 0 */
  agentTokenizedPrice: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  trendScore: number;
}

export interface InventoryStatus {
  productId: string;
  /** @minimum 0 */
  inventoryCount: number;
  available: boolean;
}

export interface DeliveryEstimate {
  productId: string;
  postalCode: string;
  estimatedFrom: string;
  estimatedTo: string;
}

export type CompatibilityCheckInputRequirements = {[key: string]: string};

export interface CompatibilityCheckInput {
  productId: string;
  requirements: CompatibilityCheckInputRequirements;
}

export interface CompatibilityResult {
  productId: string;
  compatible: boolean;
  matches: string[];
  mismatches: string[];
}

export interface CheckoutItem {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export type CheckoutInputConsumer = typeof CheckoutInputConsumer[keyof typeof CheckoutInputConsumer];


export const CheckoutInputConsumer = {
  human: 'human',
  agent: 'agent',
} as const;

export interface CheckoutInput {
  consumer: CheckoutInputConsumer;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  paymentToken?: string | null;
  /** @minItems 1 */
  items: CheckoutItem[];
}

export type CheckoutResponseConsumer = typeof CheckoutResponseConsumer[keyof typeof CheckoutResponseConsumer];


export const CheckoutResponseConsumer = {
  human: 'human',
  agent: 'agent',
} as const;

export type CheckoutResponseStatus = typeof CheckoutResponseStatus[keyof typeof CheckoutResponseStatus];


export const CheckoutResponseStatus = {
  payment_pending: 'payment_pending',
  paid: 'paid',
} as const;

export type CheckoutResponsePaymentGateway = typeof CheckoutResponsePaymentGateway[keyof typeof CheckoutResponsePaymentGateway];


export const CheckoutResponsePaymentGateway = {
  payfast: 'payfast',
  agent_api: 'agent_api',
} as const;

export interface CheckoutResponse {
  orderId: string;
  consumer: CheckoutResponseConsumer;
  status: CheckoutResponseStatus;
  paymentGateway: CheckoutResponsePaymentGateway;
  /** @nullable */
  redirectUrl?: string | null;
}

export type ListProductsParams = {
q?: string;
category?: string;
availableOnly?: boolean;
};

export type GetEstimatedDeliveryDatesParams = {
productId: string;
/**
 * @minLength 2
 * @maxLength 20
 */
postalCode: string;
};