# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutInputConsumer } from './checkoutInputConsumer';
import type { CheckoutItem } from './checkoutItem';

export interface CheckoutInput {
  consumer: CheckoutInputConsumer;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  paymentToken?: string | null;
  /** @minItems 1 */
  items: CheckoutItem[];
}