# lib/api-zod/src/generated/types/checkoutItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutItem {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}