# lib/api-zod/src/generated/types/checkoutResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutResponseConsumer } from './checkoutResponseConsumer';
import type { CheckoutResponsePaymentGateway } from './checkoutResponsePaymentGateway';
import type { CheckoutResponseStatus } from './checkoutResponseStatus';

export interface CheckoutResponse {
  orderId: string;
  consumer: CheckoutResponseConsumer;
  status: CheckoutResponseStatus;
  paymentGateway: CheckoutResponsePaymentGateway;
  /** @nullable */
  redirectUrl?: string | null;
}