# lib/api-zod/src/generated/types/compatibilityResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompatibilityResult {
  productId: string;
  compatible: boolean;
  matches: string[];
  mismatches: string[];
}