# lib/api-zod/src/generated/types/inventoryStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryStatus {
  productId: string;
  /** @minimum 0 */
  inventoryCount: number;
  available: boolean;
}