# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductDetailedSpecs } from './productDetailedSpecs';

export interface Product {
  id: string;
  /** @nullable */
  gtin?: string | null;
  title: string;
  detailedSpecs: ProductDetailedSpecs;
  /** @minimum 0 */
  inventoryCount: number;
  /** @minimum 0 */
  humanPrice: number;
  /** @minimum 0 */
  agentTokenizedPrice: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  trendScore: number;
}